Rename mob files
