+++ src/base/gameengine.hpp
@@ -31,8 +31,8 @@
 #include "base/entity.hpp"
 #include "map/dungeon.hpp"
 #include "map/lightmap.hpp"
-#include "mob_creature.hpp"
-#include "mob_player.hpp"
+#include "mob/creature.hpp"
+#include "mob/player.hpp"
 #include "screen.hpp"
 #include "spell_fireball.hpp"
 #include "ui_dialog.hpp"
