+++ src/mob/behavior.hpp
