+++ src/mob/friend.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009 Jice
+ * Copyright (c) 2010 Jice
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -24,31 +24,46 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 #pragma once
-#include "mob_creature.hpp"
+#include <libtcod.hpp>
 
-class Minion : public Creature {
- public:
-  Minion();
-  bool update(float elapsed) override;
-  void setSeen();
-  void onReplace() override;
+#include "mob/creature.hpp"
+#include "util_textgen.hpp"
 
- protected:
-  float pathTimer;
-  bool seen;
+enum FriendAiMode {
+  AI_CATCH_ME,
+  AI_HIDE_AND_SEEK,
 };
 
-class Villager : public Minion {
+class Friend : public Creature {
  public:
-  Villager();
+  Friend();
   bool update(float elapsed) override;
+  float getWalkCost(int xFrom, int yFrom, int xTo, int yTo, void* userData) const override;
 
- protected:
-  static float talkDelay;
-};
+  // SaveListener
+  bool loadData(uint32_t chunkId, uint32_t chunkVersion, TCODZip* zip) override;
+  void saveData(uint32_t chunkId, TCODZip* zip) override;
 
-class Archer : public Minion {
- public:
-  Archer();
-  bool update(float elapsed) override;
+ private:
+  TextGenerator* talkGenerator = nullptr;
+  float timer;
+  bool startPhrase, foodTuto, foodObj;
+
+  // catch me vars
+  float lostDelay;
+
+  int awayCount;
+  int caught;
+
+  // hide & seek vars
+  int counter;
+  bool see, tutoPause;
+  float sight;
+
+  bool updateCatchMe(float elapsed);
+  bool updateHideAndSeek(float elapsed);
+  bool updateFollow(float elapsed);
+  float getWalkCostCatchMe(int xFrom, int yFrom, int xTo, int yTo, void* userData) const;
+  float getWalkCostHideAndSeek(int xFrom, int yFrom, int xTo, int yTo, void* userData) const;
+  float getWalkCostFollow(int xFrom, int yFrom, int xTo, int yTo, void* userData) const;
 };
